Warn/help users who probably got the arguments in the wrong order:
  danw@technomage:blanche> ./blanche danw -a some-list

  WARNING: "LIST:some-list" was just added to list "danw".
  If you meant to add yourself to the list "some-list", type:
          blanche danw -d some-list  (to undo this)
          blanche some-list -a danw  (to add yourself to that list)

  danw@technomage:blanche> ./blanche danw -d some-other-list
  ./blanche:  Unable to find member some-other-list to delete from danw
  (If you were trying to remove yourself from the list "some-other-list",
  the correct command is "blanche some-other-list -d danw".)

The warning message is given when the user tries to add a list to the
list with his/her username, and the help message is given when the
user tries to delete something from the list with his/her username and
it fails.


Additional, unrelated change: make the usage note document
`-db/-database' instead of `-S/-server' (to match the moira client).
The old options are still supported, just not documented.