change this behavior:
  danw@technomage:~> blanche danw-test-1 -a danwn

  WARNING: "STRING:danwn" was just added to list "danw-test-1".
          If danw-test-1 is a mailing list, this may cause it to stop working.
          You should consider removing "STRING:danwn" from the list.

to give the `No records in database match ... while adding ...' error
since `danwn' was probably a typo. Because it only makes sense to add
a local string to a list if that string is the same as a list or a
user (otherwise, as the warning above states, you'll break the list).
Since blanche tries those cases first, if we reach the string case,
the string isn't the same as a user or list, so the user probably
doesn't want to add it to the list, so we assume it was a typo.

If the user had really wanted to add it as a string, he should have
said STRING:danwn, which will still work (although it will no longer
give the warning since it's not the sort of thing you're likely to do
accidentally).